Moved unit tests to proper location
